test: extract turn execution white-box suite
